backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
